moved auth into subfolder of services
